+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.3
+
+* Adds `trackingMode` to Apple Maps
+* Adds `liteModeEnabled` to Google Maps
+* Adds `mapToolbarEnabled` to Google Maps
+
 ## 1.0.2
 
 * Updates apple_maps_flutter to apply memory leak fix